backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
contourpy==1.3.1
cryptography==45.0.5
cycler==0.12.1
decorator==5.2.1
distlib==0.3.9
Django==5.2.3
django-autoslug==1.9.9
django-cors-headers==4.7.0
django-filter==25.1
django-tinymce==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.26.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
kiwisolver==1.4.8
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
networkx==3.4.2
numpy==2.2.4
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyjokes==0.8.3
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tenacity==8.5.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.29.3
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
